requirements.txt
imaplib2
email
gspread
paho-mqtt
oauth2client
gtts
pygame
python-dotenv
